example page will not be deleted

+++ multikulti/client_listener.py
@@ -25,9 +25,10 @@ def delete_old():
     to_delete = []
     for keyid in query_db("SELECT jid FROM user_queue WHERE \
             status_init<=strftime('%s',date('now',?))", [days]):
-        to_delete.append(keyid[0])
-        pat_k = path.join(app.config['USERJOB_DIRECTORY'], keyid[0])
-        rmtree(pat_k)
+        if keyid[0] != app.config['EXAMPLE_JOB']:
+            to_delete.append(keyid[0])
+            pat_k = path.join(app.config['USERJOB_DIRECTORY'], keyid[0])
+            rmtree(pat_k)
     for k in to_delete:
         query_db("DELETE FROM constraints WHERE jid=?", [k], insert=True)
     query_db("DELETE FROM user_queue where jid in (select jid FROM user_queue \
